Field Capabilities as separate package
